fast_api_server/requirements.txt
fastapi
uvicorn
pydantic
typing
paho-mqtt